Move to Hyprland + waybar
I've been having issues getting screen sharing to work with sway for
quite some time. It's running out of the box with Hyprland. So are tons
of other things. While I'm missing the simplicity of sway I set up a
minimal Hyprland config with waybar that feels the same. I'll be using
that for now.