Setup neovim with lua
Been gone for a few years and neovim is now using lua over VimL. This is
an initial setup that adds a lot of functionality:
 - basic config;
 - plugins via lazy nvim;
 - lsp support;
 - auto completion;
 - co-pilot;
 - nerd tree nvim aka nvim-tree;
 - fancy fzf aka telescope;
 - syntax hihglight via treesitter;
 - some variation of the zenburn colorscheme;

I'm really happy with how responsive this is compared to the last
experience I had with heavy plugin setup in vim. LSP support is huge!
nvim now feels better to use than VS Code and even Zed. Autocomplete is
working amazing even while typing this message. Real hype!